+++ sup3r/models/wind_conditional_moments.py
@@ -33,7 +33,7 @@ class WindCondMom(AbstractWindInterface, Sup3rCondMom):
             Keyword arguments including 'training_features', 'output_features',
             'smoothed_features', 's_enhance', 't_enhance', 'smoothing'
         """
-        AbstractWindInterface.set_model_params(self, **kwargs)
+        AbstractWindInterface.set_model_params(**kwargs)
         Sup3rCondMom.set_model_params(self, **kwargs)
 
     @tf.function
